dom/media/tests/mochitest/test_TextTrackList.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=833386
-->
<head>
	<title>Test for Bug 833386 - HTMLTrackElement</title>
	<script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
	<script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
	<script type="text/javascript" src="/tests/content/html/content/test/reflect.js"></script>
	<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<div id="videoDiv">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
	SimpleTest.waitForExplicitFinish();
	var videoDiv = document.getElementById("videoDiv");
	var WebvttEnabled = undefined;
	// Verify WebVTT support
	try{
  		WebvttEnabled = SpecialPowers.getBoolPref("media.webvtt.enabled");
	}catch (ex){
		// SpecialPowers failed, perhaps because WebVTT isn't compiled in
		console.log("media.webvtt.enabled pref not found; skipping WebVTT validation");
	}
	if (WebvttEnabled !== undefined) {
	   	// Testing with the pref set to True
	   	SpecialPowers.setBoolPref("media.webvtt.enabled", true);
		var videoElement = document.createElement("video");
		videoElement.src = "/tests/content/media/test/seek.webm";
		videoDiv.appendChild(videoElement);
		var TextTrackList = videoElement.textTracks;
		is(TextTrackList.length, 0);
		videoElement.addTextTrack("subtitles", "label", "en-CA");
		is(TextTrackList.length, 1);
	   	is(TextTrackList[0].kind, "subtitles");
		is(TextTrackList[0].label, "label");
		is(TextTrackList[0].language, "en-CA");
		is(TextTrackList[0].mode, "hidden");
		videoDiv.removeChild(videoElement);
		// Testing with the pref set to False
	   	SpecialPowers.setBoolPref("media.webvtt.enabled", false);
		var videoElement = document.createElement("video");
		videoElement.src = "/tests/content/media/test/seek.webm";
		videoDiv.appendChild(videoElement);
		var TextTrackList = videoElement.textTracks;
		is(TextTrackList.length, undefined);
		videoElement.addTextTrack("subtitles", "label", "en-CA");
		is(TextTrackList.length, undefined);
	   	is(TextTrackList[0].kind, undefined);
		is(TextTrackList[0].label, undefined);
		is(TextTrackList[0].language, undefined);
		is(TextTrackList[0].mode, undefined);
		SimpleTest.finish();
	 }else{
		var videoElement = document.createElement("video");
		videoElement.src = "/tests/content/media/test/seek.webm";
		videoDiv.appendChild(videoElement);
		var TextTrackList = videoElement.textTracks;
		is(TextTrackList.length, undefined);
		videoElement.addTextTrack("subtitles", "label", "en-CA");
		is(TextTrackList.length, undefined);
	   	is(TextTrackList[0].kind, undefined);
		is(TextTrackList[0].label, undefined);
		is(TextTrackList[0].language, undefined);
		is(TextTrackList[0].mode, undefined);
		SimpleTest.finish();
	}
</script>
</pre>	
</body>
</html>
